Move client out of index.ts because it is not meant to be public and set common library version to adding allocation key
