--- docs/node-server-sdk.epposerverclient._constructor_.md
@@ -1,21 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@eppo/node-server-sdk](./node-server-sdk.md) &gt; [EppoServerClient](./node-server-sdk.epposerverclient.md) &gt; [(constructor)](./node-server-sdk.epposerverclient._constructor_.md)
-
-## EppoServerClient.(constructor)
-
-Constructs a new instance of the `EppoServerClient` class
-
-<b>Signature:</b>
-
-```typescript
-constructor(configurationStore: IConfigurationStore, poller: IPoller);
-```
-
-## Parameters
-
-|  Parameter | Type | Description |
-|  --- | --- | --- |
-|  configurationStore | IConfigurationStore |  |
-|  poller | IPoller |  |
-
