--- docs/node-server-sdk.epposerverclient.md
@@ -1,27 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@eppo/node-server-sdk](./node-server-sdk.md) &gt; [EppoServerClient](./node-server-sdk.epposerverclient.md)
-
-## EppoServerClient class
-
-<b>Signature:</b>
-
-```typescript
-export declare class EppoServerClient extends EppoClient implements IEppoClient 
-```
-<b>Extends:</b> EppoClient
-
-<b>Implements:</b> IEppoClient
-
-## Constructors
-
-|  Constructor | Modifiers | Description |
-|  --- | --- | --- |
-|  [(constructor)(configurationStore, poller)](./node-server-sdk.epposerverclient._constructor_.md) |  | Constructs a new instance of the <code>EppoServerClient</code> class |
-
-## Methods
-
-|  Method | Modifiers | Description |
-|  --- | --- | --- |
-|  [stopPolling()](./node-server-sdk.epposerverclient.stoppolling.md) |  |  |
-
