--- docs/node-server-sdk.epposerverclient.stoppolling.md
@@ -1,15 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@eppo/node-server-sdk](./node-server-sdk.md) &gt; [EppoServerClient](./node-server-sdk.epposerverclient.md) &gt; [stopPolling](./node-server-sdk.epposerverclient.stoppolling.md)
-
-## EppoServerClient.stopPolling() method
-
-<b>Signature:</b>
-
-```typescript
-stopPolling(): void;
-```
-<b>Returns:</b>
-
-void
-
