+++ package.json
@@ -32,7 +32,7 @@
     "axios": "^0.27.2",
     "http-status-codes": "^2.2.0",
     "lru-cache": "^9.1.1",
-    "@eppo/js-client-sdk-common": "^1.5.0"
+    "@eppo/js-client-sdk-common": "^1.4.1"
   },
   "devDependencies": {
     "@google-cloud/storage": "^6.9.3",
